rename to vocab dir
