+++ src/lib/ui/Chat.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
 	import {browser} from '$app/env';
 
-	import type {Space} from '$lib/spaces/space.js';
-	import type {Member} from '$lib/members/member.js';
+	import type {Space} from '$lib/vocab/space/space.js';
+	import type {Member} from '$lib/vocab/member/member.js';
 	import ChatItems from '$lib/ui/ChatItems.svelte';
 	import {get_app} from '$lib/ui/app';
