+++ src/lib/ui/SpaceInput.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import Modal from '$lib/ui/Modal.svelte';
 	import Markup from '@feltcoop/felt/ui/Markup.svelte';
-	import type {CommunityModel} from '$lib/communities/community.js';
+	import type {CommunityModel} from '$lib/vocab/community/community.js';
 	import {autofocus} from '$lib/ui/actions';
 	import {get_app} from '$lib/ui/app';
