+++ src/lib/ui/Voice.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import type {Space} from '$lib/spaces/space.js';
+	import type {Space} from '$lib/vocab/space/space.js';
 
 	export let space: Space;
 </script>
