+++ src/lib/vocab/community/community.ts
@@ -1,5 +1,5 @@
-import type {Space} from '$lib/spaces/space.js';
-import type {Member} from '$lib/members/member.js';
+import type {Space} from '$lib/vocab/space/space.js';
+import type {Member} from '$lib/vocab/member/member.js';
 
 export interface Community {
 	community_id: number;
