+++ src/lib/vocab/persona/persona.ts
