+++ src/lib/vocab/space/space.ts
